requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pyOpenSSL==22.1.0
python-dotenv==0.21.0
redis==4.4.0
requests==2.28.1
six==1.16.0
spotipy==2.22.0
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.2.2
